+++ CHANGES.rst
@@ -7,6 +7,7 @@ Changelog
 - #2593 Always do a full reindex of the object after transition
 - #2593 Retrieve and store object's raw data on snapshot creation
 - #2593 Skip permissions update after transition when InternalUse is unchanged
+- #2595 Move sample analyses field logic to data manager
 - #2594 Fix Traceback when running migration again for Suppliers
 - #2591 Remove CreationDate index from catalogs in favour of created
 - #2592 Remove Action Handler Pool